flytter henting av TMU under Enkel fane
